feat: extract shared functionality into core package
