+++ packages/core/src/constants.ts
@@ -1,0 +1,12 @@
+export const PUBLICATION_URL = 'https://download-a.akamaihd.net'
+export const CATALOG_URL = PUBLICATION_URL + '/meps/jwl/current/catalogs/v3/catalog.db.gz'
+export const MEDIA_URL = 'https://api.hag27.com/GETPUBMEDIALINKS'
+
+export const PUBLICATION_TYPES = {
+  WATCHTOWER: 14,
+  OCLM: 30
+}
+
+export const PUBLICATION_CLASSES = {
+  WATCHTOWER_ARTICLE: 40
+}
