+++ packages/express/package.json
@@ -29,17 +29,15 @@
     "express"
   ],
   "dependencies": {
+    "@library-api/core": "0.1.0",
     "express": "^4.17.1",
     "node-fetch": "^2.6.1",
-    "sqlite": "^4.0.19",
-    "sqlite3": "^5.0.2",
     "unzipper": "^0.10.11"
   },
   "devDependencies": {
     "@types/express": "^4.17.11",
     "@types/jest": "^26.0.20",
     "@types/node-fetch": "^2.5.8",
-    "@types/sqlite3": "^3.1.7",
     "@types/unzipper": "^0.10.3",
     "jest": "^26.6.3",
     "ts-jest": "^26.4.4",
